Add summary functionality

+++ src/transaction.py
@@ -9,7 +9,8 @@
 
 class Transaction:
 
-    def __init__(self, transaction: str, category: str, amount: int, date: str):
+    def __init__(self, transaction: str, category: str, amount: float,
+                 date: str):
         self.transaction_type = transaction
         self.category = category
         self.amount = amount
